Renames. Adjust router unit test.
