--- ui/test/config.spec.ts
@@ -1,22 +1,0 @@
-import { expect } from "chai";
-import { RouterConfigurationStub } from "./stubs";
-
-// import { Config } from "../src/app/config";
-
-// tslint:disable:no-unused-expression
-
-describe("Config test suite", () => {
-    // TODO: Unit tests for our Router class
-    /*
-    it("should initialize correctly", () => {
-        const routerConfiguration = new RouterConfigurationStub();
-
-        const sut = new Config();
-
-        sut.configureRouter(routerConfiguration);
-
-        expect(routerConfiguration.title).to.equal("Hello World");
-        expect(routerConfiguration.map.calledOnce).to.be.true;
-    });
-    */
-});
